data_scraper/requirements.txt
selenium>=4.15.0
undetected-chromedriver>=3.5.4
beautifulsoup4>=4.12.0
requests>=2.31.0
Pillow>=10.1.0
pandas>=2.1.0
python-dotenv>=1.0.0
lxml>=4.9.3
tqdm>=4.66.0
easyocr>=1.7.0
